Add XF86VidModeAddModeLine man page
The function was actually already described in the man page, and
even had a commented out prototype in the man page until it was
removed by the cleanup in commit aac92ef0f20f8.

So this commit restores the prototype (but doesn't comment it out as
before), adds the function to the NAME section, and adds a .so for it.

Signed-off-by: Alan Coopersmith <alan.coopersmith@oracle.com>